ShowsDiscoverLiveData: extract local shows marker.
