+++ app/src/main/java/com/battlelancer/seriesguide/ui/search/ShowsDiscoverLiveData.kt
@@ -11,7 +11,6 @@ import com.battlelancer.seriesguide.settings.DisplaySettings
 import com.battlelancer.seriesguide.thetvdbapi.TvdbException
 import com.battlelancer.seriesguide.tmdbapi.TmdbTools2
 import com.battlelancer.seriesguide.traktapi.SgTrakt
-import com.battlelancer.seriesguide.ui.shows.ShowTools
 import com.battlelancer.seriesguide.util.Errors
 import com.uwetrottmann.androidutils.AndroidUtils
 import com.uwetrottmann.tmdb2.entities.TmdbDate
@@ -131,7 +130,7 @@ class ShowsDiscoverLiveData(val context: Context) : LiveData<ShowsDiscoverLiveDa
                     }
                 }
             }
-            markLocalShows(searchResults)
+            SearchTools().markLocalShowsAsAddedAndSetPosterPath(context, searchResults)
             return buildResultSuccess(searchResults, R.string.add_empty, false)
         }
 
@@ -145,31 +144,13 @@ class ShowsDiscoverLiveData(val context: Context) : LiveData<ShowsDiscoverLiveDa
                 } else {
                     tvdbTools.searchSeries(query, language)
                 }
-                markLocalShows(results)
+                SearchTools().markLocalShowsAsAddedAndSetPosterPath(context, results)
                 return buildResultSuccess(results, R.string.no_results, true)
             } catch (e: TvdbException) {
                 Timber.e(e, "Searching show failed")
             }
 
             return buildResultFailure(R.string.tvdb, true)
-        }
-
-        private fun markLocalShows(results: List<SearchResult>?) {
-            val existingPosterPaths = ShowTools.getSmallPostersByTvdbId(context)
-            if (existingPosterPaths == null || results == null) {
-                return
-            }
-
-            for (result in results) {
-                result.overview = String.format("(%s) %s", result.language, result.overview)
-
-                if (existingPosterPaths.indexOfKey(result.tvdbid) >= 0) {
-                    // is already in local database
-                    result.state = SearchResult.STATE_ADDED
-                    // use the poster we fetched for it (or null if there is none)
-                    result.posterPath = existingPosterPaths[result.tvdbid]
-                }
-            }
         }
 
         private fun searchShowsOnTrakt(): Result? {
